feat: 매장 정보 조회 api
